// src/app/core/enums/user-types.enum.ts
export enum UserTypes {
  companyAgent = 'companyAgent',
  playerAgent = 'playerAgent',
  company = 'company',
  admin = 'admin',
  player = 'player',
  keFu = 'keFu',          // 客服人员
  financialStaffDepositGroup = 'financialStaffDepositGroup', // 入款组财务
  financialStaffPaymentGroup = 'financialStaffPaymentGroup', // 出款组财务
  financialStaffDirector = 'financialStaffDirector',         // 财务总监
}